Optimize round key access on riscv64.
Normally we use CELL which can be optimized by the compiler to not use memcopy,
but instead just copy the data directly into the registers (as it should).
This optimization can't be done with the RISC-V code and a memcopy is inserted.
We essentially do this optimization by hand here by using a RefCell and use
the pointer to the content inside the ASM. Since we know that there can't be
any open borrows or mutable borrows, and the ASM doesn't save the ptr anywhere,
this should be sound (without doing a runtime check).

Once the intrinsics are available, we will return to use a Cell, since the code
will then properly optimize by the compiler (and not using memcopy).