Re-implement "as_array_of_cells" until it's stable.
The compiler can optimize the code the same way even if we don't
use this pattern, but using this pattern we write down out intend
more clearly. This pattern seems safe, since "as_slice_of_cells"
is stable and does essentially the same (there are also no open
questions for #88248).